Add ammo recovery to sword

+++ Scripts/Player/Weapons/BulletIndicator.cs
@@ -3,7 +3,7 @@ using Godot;
 public partial class BulletIndicator : Sprite2D
 {
 	[Export] ProgressBar BULLET_BAR;
-	public void ChangeBulletBarType(int maxAmmo, int currentAmmo) {
+	public void ChangeBulletBarType(int maxAmmo, float currentAmmo) {
 		BULLET_BAR.MaxValue = maxAmmo; BULLET_BAR.Value = currentAmmo;
 	}
 }
